[key reuse] improve some key reuse errors.
